Changed the conversion function so that it takes source and sink URI's as arguments.
Some things to figure out:

1.Do we need the option type parameter for sink UURI's just for the publish case?

In the serviceName  RESPONSE example- 403AB is truncate to 3AB, do we need to carry out truncation? Becuase from the examples provided to us in the spec, the ue_type is 403AB, but if i enter that in the test, it shows overflow, so do we have to alter our code for those cases?

Do we need to check for the fact that if publish sink is provided, it must be equal to the authority name?